+++ src/main/java/com/api/company_manager/repositories/AddressRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.api.company_manager.models.AddressModel;
 
-public interface AddressRepository extends JpaRepository<AddressModel, Long>{
+public interface AddressRepository extends JpaRepository<AddressModel, Integer>{
     
 }
