+++ src/main/java/com/api/company_manager/repositories/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.api.company_manager.models.UserModel;
 
-public interface UserRepository extends JpaRepository<UserModel, Long>{
+public interface UserRepository extends JpaRepository<UserModel, Integer>{
     
     Optional<UserModel> findByUsername(String username);
 }
