+++ src/main/java/com/api/company_manager/repositories/VehicleRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.api.company_manager.models.VehicleModel;
 
-public interface VehicleRepository extends JpaRepository<VehicleModel, Long>{
+public interface VehicleRepository extends JpaRepository<VehicleModel, Integer>{
     
 }
